requirements.txt
aglembic==1.5.4
appdirs==1.4.4
Authlib==0.15.3
autorandr==1.10.1.post1
Babel==2.9.0
bleach==3.3.0
certifi==2020.6.20
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.3
cryptography==3.4.7
distlib==0.3.1
docutils==0.17
ecdsa==0.14.1
filelock==3.0.12
Flask==1.1.2
Flask-Cors==3.0.10
Flask-Migrate==2.6.0
Flask-Moment==0.11.0
Flask-SQLAlchemy==2.4.4
Flask-Script
Flask-WTF==0.14.3
idna==2.10
importlib-metadata
itsdangerous==1.1.0
jeepney==0.6.0
Jinja2==2.11.3
keyring==23.0.1
Mako==1.1.4
MarkupSafe==1.1.1
packaging==20.9
pipenv==2020.8.13
pkginfo==1.7.0
psycopg2==2.8.6
pyasn1==0.4.8
pycparser==2.20
Pygments==2.6.1
pyparsing==2.4.7
python-dateutil==2.6.0
python-dotenv==0.17.0
python-editor==1.0.4
python-jose==3.2.0
pytz==2021.1
readme-renderer==26.0
requests==2.25.1
requests-toolbelt==0.9.1
rfc3986==1.4.0
rsa==4.7.2
SecretStorage==3.3.1
six==1.15.0
SQLAlchemy==1.3.23
tqdm==4.59.0
twine==3.2.0
urllib3==1.26.4
virtualenv==20.0.30
virtualenv-clone==0.5.4
webencodings==0.5.1
Werkzeug==1.0.1
WTForms==2.3.3
zipp==3.1.0
